backend/requirements.txt
langchain_aws
langchain
python-dotenv
flask
pyngrok
flask_cors
moviepy==1.0.3
beautifulsoup4
selenium
openai
pyPDF2